requirements.txt
asgiref==3.7.2
Django==4.2.2
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-yasg==1.21.6
inflection==0.5.1
packaging==23.1
Pillow==9.5.0
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
